# perf: faster element deletion during matching