`hasattr` is slow.


M3 Macbook
`PROFILE=0 python test/external/external_benchmark_schedule.py`
Before:
```
***** model forward in  16.69 ms
***** model schedule in   5.56 ms
***** model lower in 892.72 ms
```

After:
```
***** model forward in  16.70 ms
***** model schedule in   5.57 ms
***** model lower in 745.73 ms
```